% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_optimism.R
\name{boot_optimism}
\alias{boot_optimism}
\alias{boot_optimism.add_dummies}
\title{Bootstrap-Based Optimism}
\usage{
boot_optimism(object, ...)

\method{boot_optimism}{add_dummies}(
  object,
  mc.cores = switch(.Platform$OS.type, windows = 1L, detectCores()),
  ...
)
}
\arguments{
\item{object}{the apparent model}

\item{...}{additional parameters of function \link{boot_rule}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}}
}
\value{
Function \link{boot_optimism.add_dummies} returns an \eqn{R\times k} \link[base]{double} \link[base]{matrix}.
}
\description{
Bootstrap-Based Optimism
}
\details{
Bootstrap-based optimism
of the dichotomized predictors. Specifically,

\enumerate{

\item{\eqn{R} copies of bootstrap samples are generated. In the \eqn{j}-th bootstrap sample,
\enumerate{
\item{obtain the dichotomizing rules \eqn{\mathbf{\mathcal{D}}^{(j)}} of predictors \eqn{x_1^{(j)},\cdots,x_k^{(j)}} based on response \eqn{y^{(j)}} (via \link{boot_rule})}
\item{multivariable regression (with additional predictors \eqn{z^{(j)}}, if any) coefficient estimates \eqn{\mathbf{\hat{\beta}}^{(j)} = \left(\hat{\beta}_1^{(j)},\cdots,\hat{\beta}_k^{(j)}\right)^t} of
the dichotomized predictors \eqn{\left(\tilde{x}_1^{(j)},\cdots,\tilde{x}_k^{(j)}\right) = \mathcal{D}^{(j)}\left(x_1^{(j)},\cdots,x_k^{(j)}\right)} (via \link{add_dummies})
are the \emph{\strong{bootstrap performance estimate}}.}
}
}

\item{Dichotomize \eqn{x_1,\cdots,x_k} in the \emph{entire data} using each of the bootstrap rules \eqn{\mathcal{D}^{(1)},\cdots,\mathcal{D}^{(R)}}.
Multivariable regression (with additional predictors \eqn{z}, if any) coefficient estimates \eqn{\mathbf{\hat{\beta}}^{[j]} = \left(\hat{\beta}_1^{[j]},\cdots,\hat{\beta}_k^{[j]}\right)^t} of
the dichotomized predictors \eqn{\left(\tilde{x}_1^{[j]},\cdots,\tilde{x}_k^{[j]}\right) = \mathcal{D}^{(j)}\left(x_1,\cdots,x_k\right)} (via \link{add_dummies})
are the \emph{\strong{test performance estimate}}.}

\item{Difference between the bootstrap and test performance estimates,
an \eqn{R\times k} \link[base]{matrix} of \eqn{\left(\mathbf{\hat{\beta}}^{(1)},\cdots,\mathbf{\hat{\beta}}^{(R)}\right)} minus
another \eqn{R\times k} \link[base]{matrix} of \eqn{\left(\mathbf{\hat{\beta}}^{[1]},\cdots,\mathbf{\hat{\beta}}^{[R]}\right)},
are the \emph{\strong{bootstrap-based optimism}}.}

}
}
\references{
Ewout W. Steyerberg (2009) Clinical Prediction Models.
\doi{10.1007/978-0-387-77244-8}

Frank E. Harrell Jr., Kerry L. Lee, Daniel B. Mark. (1996)
Multivariable prognostic models: issues in developing models, evaluating
assumptions and adequacy, and measuring and reducing errors.
\doi{10.1002/(SICI)1097-0258(19960229)15:4<361::AID-SIM168>3.0.CO;2-4}
}
