% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_rule.R
\name{boot_rule}
\alias{boot_rule}
\alias{boot_rule.add_dummies}
\title{Bootstrap-Based Dichotoming Rules}
\usage{
boot_rule(object, R, ...)

\method{boot_rule}{add_dummies}(
  object,
  R = 1000L,
  mc.cores = switch(.Platform$OS.type, windows = 1L, detectCores()),
  ...
)
}
\arguments{
\item{object}{the apparent model}

\item{R}{positive \link[base]{integer} scalar,
number of bootstrap replicates \eqn{R}}

\item{...}{additional parameters, currently no use}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}}
}
\value{
Function \link{boot_rule.add_dummies} returns a \link[base]{length}-\eqn{R} \link[base]{list}
of \verb{???}
}
\description{
Bootstrap-Based Dichotoming Rules
}
\details{
Bootstrap-based optimism
of the dichotomized predictors. Specifically,

\enumerate{

\item{\eqn{R} copies of bootstrap samples are generated. In the \eqn{j}-th bootstrap sample,
\enumerate{
\item{obtain the dichotomizing rules \eqn{\mathbf{\mathcal{D}}^{(j)}} of predictors \eqn{x_1^{(j)},\cdots,x_k^{(j)}} based on response \eqn{y^{(j)}} (via \link[maxEff]{rpart1})}
}
}

}
}
\references{
Ewout W. Steyerberg (2009) Clinical Prediction Models.
\doi{10.1007/978-0-387-77244-8}

Frank E. Harrell Jr., Kerry L. Lee, Daniel B. Mark. (1996)
Multivariable prognostic models: issues in developing models, evaluating
assumptions and adequacy, and measuring and reducing errors.
\doi{10.1002/(SICI)1097-0258(19960229)15:4<361::AID-SIM168>3.0.CO;2-4}
}
