% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummies.R
\name{print.add_dummies}
\alias{print.add_dummies}
\title{print.add_dummies}
\usage{
\method{print}{add_dummies}(x, ...)
}
\arguments{
\item{x}{an \link{add_dummies} object}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{print.add_dummies} does not have a returned value
}
\description{
print.add_dummies
}
\details{
..
}
\keyword{internal}
