% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart_node1.R
\name{rpart_node1}
\alias{rpart_node1}
\title{Batch Operation of Functions \link[rpart]{rpart} and \link[maxEff]{node1}}
\usage{
rpart_node1(X, y, mc.cores = getOption("mc.cores"), ...)
}
\arguments{
\item{X}{\link[base]{data.frame} or \link[base]{list}}

\item{y}{\link[base]{vector}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}}

\item{...}{additional parameter, currently of no use}
}
\value{
Function \code{\link[=rpart_node1]{rpart_node1()}} returns a \link[stats]{listof} function \link[maxEff]{node1} returns.
}
\description{
Batch Operation of Functions \link[rpart]{rpart} and \link[maxEff]{node1}
}
\keyword{internal}
