% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummies.R
\name{add_dummies}
\alias{add_dummies}
\title{add_dummies}
\usage{
add_dummies(
  start.model,
  formula,
  data,
  rule = rpart1.(y = data[[terms(start.model)[[2L]]]], X = X),
  ...
)
}
\arguments{
\item{start.model}{a regression model,
currently \link[stats]{lm}, \link[stats]{glm} or \link[survival]{coxph} models are supported}

\item{formula}{\link[stats]{formula},
additional \link[base]{numeric} covariates \eqn{x_1,\cdots,x_k}
to be dichotomized}

\item{data}{(optional) \link[base]{data.frame}}

\item{rule}{..}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \link{add_dummies} returns an object of class \code{'add_dummies'}.
}
\description{
add_dummies
}
\details{
First, obtain the dichotomizing rules \eqn{\mathbf{\mathcal{D}}} of predictors \eqn{x_1,\cdots,x_k} based on
response \eqn{y} (via \link[maxEff]{rpart1}).

Then, \link[stats]{update} previous multivariable regression \code{start.model}
with dichotomized predictors \eqn{\left(\tilde{x}_1,\cdots,\tilde{x}_k\right) = \mathcal{D}\left(x_1,\cdots,x_k\right)}.
}
\note{
Different from \code{?maxEff::add_dummy}.  Function \link{add_dummies} adds \emph{multiple} dichotomized predictors.
}
\examples{
# see ?boc
}
