% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummies.R
\name{rpart1.}
\alias{rpart1.}
\title{Batch Operation of Function \code{\link[=rpart1]{rpart1()}}}
\usage{
rpart1.(
  X,
  mc.cores = switch(.Platform$OS.type, windows = 1L, detectCores()),
  ...
)
}
\arguments{
\item{X}{\link[base]{data.frame} or \link[base]{list}}

\item{mc.cores}{\link[base]{integer} scalar, see function \link[parallel]{mclapply}}

\item{...}{additional parameters of function \code{\link[=rpart1]{rpart1()}}}
}
\value{
Function \code{\link[=rpart1.]{rpart1.()}} returns a \link[stats]{listof} function \code{\link[=rpart1]{rpart1()}} returns.
}
\description{
Batch Operation of Function \code{\link[=rpart1]{rpart1()}}
}
\keyword{internal}
